tech: add empty release notes case handler and fix release name
