+++ VKUI/auto-update-release-notes/src/parsing/parsePullRequestReleaseNotesBody.ts
@@ -1,0 +1,17 @@
+import { ReleaseNoteData } from '../types';
+import { releaseNotesUpdater } from './releaseNotesUpdater';
+
+export function parsePullRequestReleaseNotesBody(
+  releaseNotesBody: string,
+  prNumber: number,
+): ReleaseNoteData[] | null | '' {
+  const updater = releaseNotesUpdater(releaseNotesBody);
+
+  return updater.getReleaseNotesData().map((change) => ({
+    ...change,
+    data: change.data.map((item) => ({
+      ...item,
+      pullRequestNumber: prNumber,
+    })),
+  }));
+}
